Add logger

+++ src/main/java/com/godeltech/pt11/service/CarService.java
@@ -2,14 +2,11 @@ package com.godeltech.pt11.service;
 
 import com.godeltech.pt11.converter.CarMapper;
 import com.godeltech.pt11.dto.CarDTO;
+import com.godeltech.pt11.entity.Car;
 import com.godeltech.pt11.entity.enums.Colour;
 import com.godeltech.pt11.exceptions.CarNotFoundException;
 import com.godeltech.pt11.repository.CarRepository;
-import com.godeltech.pt11.rest.CarController;
 import com.google.common.collect.Lists;
-import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,14 +14,13 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.stream.Collectors;
 
 @Transactional
-@RequiredArgsConstructor
 @Service
 public class CarService {
 
-    private static Logger logger = LoggerFactory.getLogger(CarService.class);
-
+    @Autowired
     private CarRepository carRepository;
 
+    @Autowired
     private CarMapper carMapper;
 
     public Iterable<CarDTO> getAllCars() {
@@ -35,11 +31,14 @@ public class CarService {
     }
 
     public void deleteCar(Long id) {
-        carRepository.deleteById(id);
+        if (carRepository.findById(id).isPresent()) {
+            carRepository.deleteById(id);
+        } else {
+            throw new CarNotFoundException(id);
+        }
     }
 
     public CarDTO createCar(CarDTO carDTO) {
-        logger.info("Creating new Car");
         return carMapper
                 .fromEntity(carRepository
                         .save(carMapper
@@ -53,15 +52,15 @@ public class CarService {
                         .orElseThrow(() -> new CarNotFoundException(id)));
     }
 
-    public CarDTO updateCar(CarDTO carDTO) {
-        Long carId = carDTO.getCarId();
-        if (carRepository.findById(carId).isPresent()) {
+    public CarDTO updateCar(CarDTO carDTO, Long id) {
+        Car car = carMapper.toEntity(carDTO);
+        car.setCarId(id);
+        if (carRepository.findById(id).isPresent()) {
             return carMapper
                     .fromEntity(carRepository
-                            .save(carMapper.toEntity(carDTO)));
+                            .save(car));
         } else {
-            logger.error("Car with " + carId + " not exists in database");
-            throw new CarNotFoundException(carId);
+            throw new CarNotFoundException(id);
         }
     }
